+++ theme-lipstick.go
@@ -103,6 +103,8 @@ func LipstickTheme() *Theme {
 		NewStyle("styled/h4").WithFont("bold"),
 		NewStyle("styled/pre").WithBackground("$bg0"),
 		NewStyle("styled/code").WithForeground("$fuchsia"),
+		NewStyle("styled/bq").WithColors("$fg2", "$bg1"),
+		NewStyle("styled/hr").WithForeground("$gray"),
 		NewStyle("switcher"),
 		NewStyle("table").WithColors("", "").WithBorder("thin $fg2"),
 		NewStyle("table:focused").WithBorder("double $fuchsia"),