+++ theme-midnight-neon.go
@@ -93,6 +93,8 @@ func MidnightNeonTheme() *Theme {
 		NewStyle("styled/h4").WithColors("$fg0", "$bg1").WithFont("bold"),
 		NewStyle("styled/pre").WithBackground("$bg0"),
 		NewStyle("styled/code").WithForeground("$aqua"),
+		NewStyle("styled/bq").WithColors("$fg2", "$bg1"),
+		NewStyle("styled/hr").WithForeground("$fg3"),
 		NewStyle("switcher"),
 		NewStyle("table").WithColors("", "").WithBorder("thin $fg3"),
 		NewStyle("table:focused").WithBorder("double $aqua"),